Fix deleting old DB entries
The keys are based on the parent inode and the filename, not the file's
inode, so the unused entries are those whose filename we didn't run
into, not whose inode we didn't run into.

Use hash tables from glib due to laziness and the fact that using plain
arrays with strings would likely slow things down too much already.